# src/Store/Web/Client/OpenAPIs/swagger.yaml
{
...
    "openapi": "3.0.0",
    "info": {
        "title": "My Title",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:7188"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "operationId": "Cart_GetCart",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/api/cart/items": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "operationId": "Cart_AddCartItem",
                "requestBody": {
                    "x-name": "request",
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCartItemRequest"
                            }
                        }
                    },
                    "required": true,
                    "x-position": 1
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/api/cart/items/{cartItemId}/quantity": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "operationId": "Cart_UpdateCartItemQuantity",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        },
                        "x-position": 1
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "x-position": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartItem"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/api/cart/items/{cartItemId}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "operationId": "Cart_RemoveCartItem",
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        },
                        "x-position": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cart": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    }
                }
            },
            "CartItem": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "productId": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "regularPrice": {
                        "type": "number",
                        "format": "decimal",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number",
                        "format": "double"
                    },
                    "total": {
                        "type": "number",
                        "format": "decimal"
                    }
                }
            },
            "AddCartItemRequest": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "productId": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "regularPrice": {
                        "type": "number",
                        "format": "decimal",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        }
    }
}